move 'merge_directories' to conan.util.files
