+++ conans/test/unittests/util/files/merge_directories_test.py
@@ -5,9 +5,8 @@ import unittest
 from os.path import join
 
 from conans import load
-from conans.client.source import merge_directories
 from conans.test.utils.test_files import temp_folder
-from conans.util.files import mkdir, save
+from conans.util.files import mkdir, save, merge_directories
 
 
 class MergeDirectoriesTest(unittest.TestCase):
